ventas e ingreso listo

+++ resources/views/admin/ingreso/create.blade.php
@@ -85,6 +85,10 @@
                             <input type="number" name="costoventa" id= "costoventa"  min="0.1" step="0.01" class="form-control borde required" required readonly />
                         </div>
                         </div>
+                        <div class="col-md-6 mb-3">
+                            <label class="form-label is-required">FACTURA PAGADA</label>
+                            <input type="text" name="pagada" id= "pagada"  class="form-control borde " required readonly />
+                       </div>
                         <div class="col-md-12 mb-3">
                              <label class="form-label">OBSERVACION</label>
                             <input type="text" name="observacion" id="observacion" class="form-control borde" />
@@ -103,21 +107,21 @@
                         </div>
                         <div class="col-md-6 mb-3">
                              <label class="form-label" id="labelcantidad">CANTIDAD</label>
-                            <input type="number" name="cantidad" id="cantidad" class="form-control borde" />
+                            <input type="number" name="cantidad" id="cantidad" min="1" step="1" class="form-control borde" />
                             @error('cantidad') <small class="text-danger">{{$message}}</small> @enderror
                         </div>
                         <div class="col-md-4 mb-3">
                             <div class="input-group">
                             <label class="form-label input-group"  id="labelpreciounitarioref">PRECIO UNITARIO (REFERENCIAL):</label>
                             <span class="input-group-text" id="spanpreciounitarioref"></span> 
-                            <input type="number" name="preciounitario" min="0.1" step="0.01" id="preciounitario" readonly class="form-control borde" />
+                            <input type="number" name="preciounitario" min="0" step="0.01" id="preciounitario" readonly class="form-control borde" />
                         </div>
                         </div> 
                         <div class="col-md-4 mb-3">
                             <div class="input-group">
                              <label class="form-label input-group" id="labelpreciounitario">PRECIO UNITARIO:</label>
                              <span class="input-group-text" id="spanpreciounitario"></span> 
-                             <input type="number" name="preciounitariomo" min="0.1" step="0.01" id="preciounitariomo" class="form-control borde" />
+                             <input type="number" name="preciounitariomo" min="0" step="0.01" id="preciounitariomo" class="form-control borde" />
                             </div>
                         </div>
                         <div class="col-md-4 mb-3">
@@ -131,15 +135,20 @@
                             <div class="input-group">
                              <label class="form-label input-group" id="labelpreciototal">PRECIO TOTAL POR PRODUCTO:</label>
                              <span class="input-group-text" id="spanpreciototal"></span>
-                            <input type="number" name="preciofinal" min="0.1" step="0.01" id="preciofinal" readonly class="form-control borde" />
+                            <input type="number" name="preciofinal" min="0" step="0.01" id="preciofinal" readonly class="form-control borde" />
                             </div>
                         </div>
+                        <div class="col-md-8 mb-3"> 
+                            <label class="form-label " id="labelobservacionproducto">OBSERVACION(Nro Serie):</label>
+                           <input type="text" name="observacionproducto"   id="observacionproducto"  class="form-control borde gui-input" />
+                       </div>
                         <button type="button" class="btn btn-info" id="addDetalleBatch"><i class="fa fa-plus"></i> Agregar Producto a la Venta</button>
                         <div class="table-responsive">
                         <table class="table table-row-bordered gy-5 gs-5" id="detallesVenta">
                             <thead class="fw-bold text-primary">
                                 <tr>
                                     <th>PRODUCTO</th>
+                                    <th>OBSERVACION</th>
                                     <th>CANTIDAD</th>
                                     <th>PRECIO UNITARIO(REFERENCIAL)</th>
                                     <th>PRECIO UNITARIO</th>
@@ -236,24 +245,27 @@
             var servicio = $('[name="servicio"]').val();
             var preciofinal = $('[name="preciofinal"]').val();
             var preciounitariomo = $('[name="preciounitariomo"]').val();
-             
+            var observacionproducto = $('[name="observacionproducto"]').val();
             //alertas para los detallesBatch
             
             if (!product) {  alert("Seleccione un Producto"); return;   }
             if (!cantidad) {  alert("Ingrese una cantidad"); return;   }
             if (!preciounitariomo) {  alert("Ingrese una cantidad"); return;   }
+            if (!servicio) {  alert("Ingrese un servicio"); return;   }
+            if (!observacionproducto) {alert("ingrese una observacion(Nro Serie):");   $("#observacionproducto").focus(); return;   }
 
             
             //$("#product option:contains('Seleccione una opción')").attr('selected',false);  
             var LVenta = [];
             var tam = LVenta.length;
-            LVenta.push(product,nameproduct,cantidad,preciounitario,servicio,preciofinal,preciounitariomo);
+            LVenta.push(product,nameproduct,cantidad,preciounitario,servicio,preciofinal,preciounitariomo,observacionproducto);
         
                 filaDetalle ='<tr id="fila' + indice + 
                 '"><td><input  type="hidden" name="Lproduct[]" value="' + LVenta[0]  + '"required>'+ LVenta[1]+
+                '</td><td><input  type="hidden" name="Lobservacionproducto[]" id="observacionproducto' + indice +'" value="' + LVenta[7] + '"required>'+ LVenta[7]+
                 '</td><td><input  type="hidden" name="Lcantidad[]" id="cantidad' + indice +'" value="' + LVenta[2] + '"required>'+ LVenta[2]+
                 '</td><td><input  type="hidden" name="Lpreciounitario[]" id="preciounitario' + indice +'" value="' + LVenta[3] + '"required>'+simbolomonedaproducto+ LVenta[3]+ 
-                    '</td><td><input  type="hidden" name="Lpreciounitariomo[]" id="preciounitariomo' + indice +'" value="' + LVenta[6] + '"required>'+ simbolomonedafactura+ LVenta[6]+ 
+                '</td><td><input  type="hidden" name="Lpreciounitariomo[]" id="preciounitariomo' + indice +'" value="' + LVenta[6] + '"required>'+ simbolomonedafactura+ LVenta[6]+ 
                 '</td><td><input  type="hidden" name="Lservicio[]" id="servicio' + indice +'" value="' + LVenta[4] + '"required>'+ simbolomonedafactura+ LVenta[4]+
                 '</td><td ><input id="preciof' + indice +'"  type="hidden" name="Lpreciofinal[]" value="' + LVenta[5] + '"required>'+ simbolomonedafactura+ LVenta[5]+ 
                 '</td><td><button type="button" class="btn btn-danger" onclick="eliminarFila(' + indice + ')" data-id="0">ELIMINAR</button></td></tr>';
@@ -265,7 +277,7 @@
                 ventatotal = parseFloat(ventatotal) + parseFloat(preciototalI);
 
                 limpiarinputs();
-                document.getElementById('costoventa').value = ventatotal; 
+                document.getElementById('costoventa').value = (ventatotal.toFixed(2)); 
 
                 var funcion="agregar";
                 botonguardar(funcion);
@@ -287,30 +299,30 @@
             cant.setAttribute("max",$stock);
             cant.setAttribute("min",1);
             if($price != null){
-                preciounit = $price;
+                preciounit = ($price).toFixed(2);
                 if(monedaproducto=="dolares" && monedafactura=="dolares"){
                     simbolomonedaproducto="$";
-                    preciototalI = $price;
-                    document.getElementById('preciounitario').value = $price;
-                    document.getElementById('preciounitariomo').value = $price;
-                    document.getElementById('preciofinal').value = $price; 
+                    preciototalI = ($price).toFixed(2);
+                    document.getElementById('preciounitario').value = ($price).toFixed(2);
+                    document.getElementById('preciounitariomo').value = ($price).toFixed(2);
+                    document.getElementById('preciofinal').value = ($price).toFixed(2); 
                 }else if(monedaproducto=="soles" && monedafactura=="soles"){
                     simbolomonedaproducto="S/.";
-                    preciototalI = $price;
-                    document.getElementById('preciounitario').value = $price;
-                    document.getElementById('preciounitariomo').value = $price; 
-                    document.getElementById('preciofinal').value = $price; 
+                    preciototalI = ($price).toFixed(2);
+                    document.getElementById('preciounitario').value = ($price).toFixed(2);
+                    document.getElementById('preciounitariomo').value = ($price).toFixed(2); 
+                    document.getElementById('preciofinal').value = ($price).toFixed(2); 
                 }else if(monedaproducto=="dolares" && monedafactura=="soles"){
                     simbolomonedaproducto="$";
                     preciototalI = ($price*mitasacambio1).toFixed(2);
-                    document.getElementById('preciounitario').value = ($price);
+                    document.getElementById('preciounitario').value = ($price).toFixed(2);
                     document.getElementById('preciounitariomo').value = ($price*mitasacambio1).toFixed(2);
                     document.getElementById('preciofinal').value = ($price*mitasacambio1).toFixed(2); 
                 }
                 else if(monedaproducto=="soles" && monedafactura=="dolares"){
                     simbolomonedaproducto="S/.";
                     preciototalI = ($price/mitasacambio1).toFixed(2);;
-                    document.getElementById('preciounitario').value = ($price);
+                    document.getElementById('preciounitario').value = ($price).toFixed(2);
                     document.getElementById('preciounitariomo').value = ($price/mitasacambio1).toFixed(2);
                     document.getElementById('preciofinal').value = ($price/mitasacambio1).toFixed(2); 
                 }
@@ -345,12 +357,13 @@
                 $("#fechav").prop("required", true);
                 var fechav = document.getElementById("labelfechav");
                 fechav.className += " is-required";
-                
+                document.getElementById('pagada').value = "NO";
             } else if ($mimoneda == "contado") {
                 $("#fechav").prop("readonly", true);
                 $("#fechav").prop("required", false); 
                 var fechav = document.getElementById("labelfechav");
-                fechav.className -= " is-required";
+                fechav.className = "form-label ";
+                document.getElementById('pagada').value = "SI";
             } 
    });
     });
@@ -388,7 +401,7 @@
     $('#fila' + ind).remove();
         indice-- ;
     // damos el valor
-    document.getElementById('costoventa').value = ventatotal;
+    document.getElementById('costoventa').value = (ventatotal.toFixed(2)); 
     //alert(resta);
 
     var funcion="eliminar";
@@ -448,7 +461,7 @@ function limpiarinputs(){
     document.getElementById('preciofinal').value = "";
     document.getElementById('preciounitario').value = "";
     document.getElementById('preciounitariomo').value = "";
-   // document.getElementById('observacionproducto').value = "";
+    document.getElementById('observacionproducto').value = "";
     monedaproducto="";
     simbolomonedaproducto="";
 }
